Retire the usage of the VariableRoleAssigner.
The VariableRoleAssigner analyzes the program to figure out whether
variables are in/out parameters or local to the method. Local variables
are made to be values and composite structures passed in/out are set
as pointers. The information about what goes in/out of the method is
already there in the compiler, so keep that information around.

The code for the VRA is left in the tree for now in case I happened
to break something and we need to reintroduce it.