Update travis config to support branches
The travis config is updated to support building other branches than master
of feldspar-language.
With this change travis will clone the feldspar-language repository and
then try to check out a branch with the same name as the current branch in
feldspar-compiler. If that fails we will check out the master branch.