Replace fromCoreM with a function that takes UntypedFeld as argument
I want to make a test case with an expression that can't be written using Feldspar's front end. So I need to construct an UntypedFeld expression and generate code for it.

Previously, `reifyFeldM` was used, which gives the next variable ID as its final state, and this state was used to initialize the code generation state. Now `reifyFeld` is used instead, so the state is lost. But `fromCoreUT` recovers the state using `allVars`. This is a bit unnecessary when only considering `fromCore`, but it has the advantage that `fromCoreUT` does not depend on some state to avoid capture.