Add generation of array operations for nonflat arrays.
Until this patch, all array operations were implemented in the array
library over a generic and relatively weakly typed representation.
Arrays could either be flat, with elements not containing arrays,
or the array could have elements each of which was a single array.
An array with structs mixing flat data and arrays was not possible.

This patch lifts this limitation by generating specialized versions
of array operations for non flat arrays.