Make the structure of arrays visible in the generated code.
Before this commit, an array has been represented as a pointer to
a generic struct array, defined in the array library, independent
of their element type. This commit introduces type a specific
array representation in terms of Program level structs, making
the array representation explicit in the generated program. This
is a pre requisite for further array optimizations.